リファクタリング: 関数をファイルに切り出し
